refactor: move checkValidConfig to config.js
